requirements.txt
# Non-Docker: install PyTorch first (CUDA and/or Intel XPU build from pytorch.org — see README).
#   pip install -r requirements.txt

numpy>=1.24
pydantic>=2
langchain-core
langchain-community
langchain-huggingface
langchain-openai
langchain-text-splitters
langgraph
openai
BCEmbedding

# Hugging Face tokenizers: SentencePieceExtractor needs both (missing either → TikToken fallback → .bpe.model parse error).
sentencepiece>=0.1.99
protobuf>=3.20
pypdf
dulwich
python-docx
openpyxl
beautifulsoup4
